[chore] 안쓰는 코드, 주석 삭제
