+++ BongBaek/BongBaek/Presentation/AccountDeletionConfirmView.swift
@@ -57,5 +57,3 @@ struct AccountDeletionConfirmView: View {
 
     }
 }
-
-
