+++ BongBaek/BongBaek/Presentation/AccountDeletionView.swift
@@ -267,4 +267,3 @@ struct DeletionReasonButton: View {
         }
     }
 }
-
