+++ BongBaek/BongBaek/Presentation/Home/Schedule/ScheduleAlarmCellView.swift
@@ -83,18 +83,3 @@ struct ScheduleIndicatorCellView: View {
         .cornerRadius(10)
     }
 }
-
-extension Color {
-    init(hex: String) {
-        let hexSanitized = hex.replacingOccurrences(of: "#", with: "")
-        let scanner = Scanner(string: hexSanitized)
-        var rgb: UInt64 = 0
-        scanner.scanHexInt64(&rgb)
-
-        let r = Double((rgb & 0xFF0000) >> 16) / 255.0
-        let g = Double((rgb & 0x00FF00) >> 8) / 255.0
-        let b = Double(rgb & 0x0000FF) / 255.0
-
-        self.init(red: r, green: g, blue: b)
-    }
-}
