+++ BongBaek/BongBaek/Presentation/Recommend/Test/LoadingView.swift
@@ -17,6 +17,3 @@ struct LoadingView: View {
         }
     }
 }
-#Preview {
-    LoadingView()
-}
