+++ BongBaek/BongBaek/Presentation/Recommend/View/RecommendCostView.swift
@@ -371,11 +371,3 @@ extension NumberFormatter {
         return formatter
     }()
 }
-
-// MARK: - Preview
-#Preview {
-    @Previewable @StateObject var stepManager = GlobalStepManager()
-    
-    RecommendCostView()
-        .environmentObject(stepManager)
-}
