+++ BongBaek/BongBaek/Presentation/Recommend/View/RecommendFlow/EventInformationView.swift
@@ -224,7 +224,3 @@ struct NextButton: View {
         .animation(.easeInOut(duration: 0.2), value: isEnabled)
     }
 }
-
-#Preview {
-    EventInformationView()
-}
