+++ BongBaek/BongBaek/Presentation/Recommend/View/RecordCompleteView.swift
@@ -72,7 +72,3 @@ struct RecordCompleteView: View {
 
     }
 }
-
-#Preview {
-    RecordCompleteView()
-}
