+++ BongBaek/BongBaek/Presentation/Recommend/View/RelationshipButton.swift
@@ -40,5 +40,3 @@ struct RelationshipButton: View {
         .buttonStyle(PlainButtonStyle())
     }
 }
-
-
